+++ ChangeLog.md
@@ -4,6 +4,7 @@
 - FIX do not RegisterActionMethod if Metadata.ActionProvider.ExecutedCallback is already set
 - add ButtonActionVM. derived from ActionVM supporting common button properties
 - add some UIProperties (ObjectVM.UI)
+- FIX ListVM clear item.Parent when removed
 
 # 0.20
 - BREAKING CHANGE ValueChangedEventArgs swap arguments and rename PreviousValue to OldValue, so that you get a compiler error on event handlers. If you create ValueChangedEventArgs you have to swap the arguments manually. No Warning in this case!